+++ src/MicroElements.Metadata/Metadata/Property.SchemaBuilder.cs
@@ -10,6 +10,7 @@ namespace MicroElements.Metadata
     /// Property schema builder implementation.
     /// </summary>
     public partial class Property<T> :
+        IConfigurableBuilder<Property<T>, PropertyData<T>>,
         ISchemaBuilder<Property<T>, ISchemaDescription>,
         ISchemaBuilder<Property<T>, IDefaultValue<T>>,
         ISchemaBuilder<Property<T>, IDefaultValue>,
@@ -17,15 +18,45 @@ namespace MicroElements.Metadata
         ISchemaBuilder<Property<T>, IExamples<T>>
     {
         /// <inheritdoc />
+        public PropertyData<T> GetState()
+        {
+            Property<T> source = this;
+            return new PropertyData<T>
+            {
+                Name = source.Name,
+                Description = source.Description,
+                Alias = source.Alias,
+                DefaultValue = source.DefaultValue,
+                Examples = source.Examples,
+                Calculator = source.Calculator,
+            };
+        }
+
+        /// <inheritdoc />
+        public Property<T> With(PropertyData<T> propertyData)
+        {
+            var property = new Property<T>(
+                    name: propertyData.Name,
+                    description: propertyData.Description,
+                    defaultValue: propertyData.DefaultValue,
+                    examples: propertyData.Examples,
+                    calculator: propertyData.Calculator,
+                    alias: propertyData.Alias)
+                .SetMetadataFrom(this);
+
+            return property;
+        }
+
+        /// <inheritdoc />
         public Property<T> With(ISchemaDescription schemaDescription)
         {
-            return this.WithRewriteFast((ref PropertyData<T> data) => data.Description = schemaDescription.Description);
+            return this.WithRewriteFast((ref PropRefData<T> data) => data.Description = schemaDescription.Description);
         }
 
         /// <inheritdoc />
         public Property<T> With(IDefaultValue<T> defaultValue)
         {
-            return this.WithRewriteFast((ref PropertyData<T> data) => data.DefaultValue = defaultValue);
+            return this.WithRewriteFast((ref PropRefData<T> data) => data.DefaultValue = defaultValue);
         }
 
         /// <inheritdoc />
@@ -43,13 +74,13 @@ namespace MicroElements.Metadata
         /// <inheritdoc />
         public Property<T> With(IPropertyCalculator<T> calculator)
         {
-            return this.WithRewriteFast((ref PropertyData<T> data) => data.Calculator = calculator);
+            return this.WithRewriteFast((ref PropRefData<T> data) => data.Calculator = calculator);
         }
 
         /// <inheritdoc />
         public Property<T> With(IExamples<T> examples)
         {
-            return this.WithRewriteFast((ref PropertyData<T> data) => data.Examples = examples);
+            return this.WithRewriteFast((ref PropRefData<T> data) => data.Examples = examples);
         }
     }
 }